requirements.txt
asgiref==3.8.1
black==24.10.0
Brotli==1.1.0
cffi==1.17.1
click==8.1.7
cssselect2==0.7.0
django-cleanup==9.0.0
django-crispy-forms==2.3
Django==5.1.3
djangorestframework
djangorestframework-simplejwt
fonttools==4.54.1
google-api-python-client
google-auth
google-auth-httplib2
google-auth-oauthlib
gunicorn==23.0.0
html5lib==1.1
mypy-extensions==1.0.0
pathspec==0.12.1
Pillow<11
platformdirs==4.3.6
psycopg2-binary==2.9.10
pycparser==2.22
pydyf==0.11.0
pyphen<0.16
python-telegram-bot
six==1.16.0
sqlparse==0.5.1
tinycss2==1.4.0
tomli==2.0.1
weasyprint==63.0
webencodings==0.5.1
zopfli==0.2.3
crispy-bootstrap4
drf-yasg